+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 ### Added
 ### Fixed
+
+- fix #11: ```sh code block for command output
+
 ### Changed
 ### Deprecated
 ### Removed